# serve static assets much faster