# before

```
$ wrk http://localhost:8080/assets/app-9505931b1971f9a49a26.css -c 10 -t 2 -d 2m
Running 2m test @ http://localhost:8080/assets/app-9505931b1971f9a49a26.css
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    72.53ms   40.65ms 431.97ms   86.88%
    Req/Sec    72.85     19.52   121.00     68.59%
  17466 requests in 2.00m, 784.68MB read
Requests/sec:    145.44
Transfer/sec:      6.53MB
```

# after

```
bc@zebrina:~/steemit/condenser$ wrk http://localhost:8080/assets/app-9505931b1971f9a49a26.css -c 10 -t 2 -d 2m
Running 2m test @ http://localhost:8080/assets/app-9505931b1971f9a49a26.css
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    27.15ms   33.89ms 236.93ms   81.94%
    Req/Sec   394.31    136.71     1.14k    70.34%
  94199 requests in 2.00m, 26.95GB read
Requests/sec:    784.82
Transfer/sec:    229.91MB

```